moved self-update code to update_flash module
